==> 08_data-cleaning/02_tidyverse/analysis/02_clean_data.Rmd <==
---
title: 'Data Analysis in R: Lecture Notes'
author: "PS239T"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    df_print: paged
    theme: flatly
    highlight: pygments
---

# 0) NOTES

Purpose of document is to clean dataset of Florida police shootings from 2009 to 2014. 

See github page for more information about dataset: https://github.com/tbtimes/florida-police-shooting-data



**********************************************************

# 1) SET UP 

## 1.1. General 

```{r}
# remove all objects
rm(list=ls())

# Unload all packages 
library(pacman)
pacman::p_unload(all)

# Add packages 
pacman::p_load(
  tidyverse, #dplyr, readr, etc.
  data.table, #fread() 
  foreign, #load data types including stata .dta files 
  magrittr, #%<>% operator
  skimr, #for summerising
  lubridate #dates
)
```

## 1.2. Load Data 

```{r}
load("../data/fl_shooting_data.RData")
```

**********************************************************

# 2) CLEAN DATA

## 2.1. Check Data 

Let's remind ourselves what varibales are in the fl_shootings dataset
```{r}
fl_shootings %>% skimr::skim()
```

Note that for any varible that we use, we want to check for missing values that are not coded as missing. For example, empty cells may not be coded as missing automatically. 
```{r}
# Trim whitespace from ends of all character variables 
fl_shootings %<>%
  mutate_if(is.character, list(str_trim))

# Check for empty cells 
fl_shootings %>%
  # Keep only character variables 
  select_if(is.character) %>%
  # Recode character variables as 1 if cell is empty 
  mutate_all(list(~ifelse(.=="",1,0))) %>%
  # Add up empty cells for each character variable 
  summarise_all(sum, na.rm=T) %>%
  # Transpose data for visibility 
  t() 
```

*Note* dplyr has versions of most functions with the "_all", "_at", and "_if" suffixes. These are invaluable if you want to make the same modification to multiple variables. 

## 2.2. Gen vars with diff types 

In your data cleaning assignment, you may need to generate new variables with different data types in order to show that you know how to manipulate character, logical, numeric, integer, and factor variables. 
```{r}
# Check types & preview 
fl_shootings %>% select(suspect_mentally_ill, race, condition) %>% glimpse()

# Add/convert variables 
fl_shootings %<>%
  # Convert suspect_mentally_ill to logical type  
  mutate(suspect_mentally_ill = as.logical(toupper(suspect_mentally_ill))) %>%
  # Turn race into factor var
  mutate(race = as.factor(race)) %>% 
  # Create binary variable for suicide 
  mutate(is_suicide = ifelse(condition == "suicide", T, F))

# Check types & preview 
fl_shootings %>% select(suspect_mentally_ill, race, is_suicide) %>% glimpse()
```

## 2.3. Date variable 
```{r}
# Check types & preview 
fl_shootings %>% select(date) %>% glimpse()

# Convert/add variables 
fl_shootings %<>%
  # Convert date variable to date type
  mutate(date_orig = date,
         date = lubridate::mdy(date_orig)) %>%
  # Add year variable 
  mutate(year = year(date))

# Check types & preview 
fl_shootings %>% select(date, date_orig, year) %>% glimpse()
```

## 2.4. Aggregate variable
```{r}
# Add variable with total people shot in city in year 
fl_shootings %<>%
  group_by(city_id, year) %>%
  mutate(total_people_city_year = n() ) %>% 
  ungroup()

# Check aggregation 
fl_shootings %>% select(id, city_id, year, total_people_city_year) %>% 
  sample_n(100) %>% arrange(city_id, year)
```

## 2.5. Recode categorical variable 
```{r}
# Frequency of weapons 
fl_shootings %>% group_by(weapon_choices) %>% summarise(n=n())

# Create new weapon var
fl_shootings %<>%
  mutate(weapon_cat = NA) %>%
  mutate(weapon_cat = ifelse(weapon_choices %in% c("Firearm"), "Gun", weapon_cat),
         weapon_cat = ifelse(weapon_choices %in% c("Blade/stabbing implement"), "blade", weapon_cat),
         weapon_cat = ifelse(weapon_choices %in% c("Blunt/bludgeoning weapon"), "blunt", weapon_cat),
         weapon_cat = ifelse(weapon_choices %in% c("BB/Pellet Gun", "Taser"), "nonlethal", weapon_cat),
         weapon_cat = ifelse(weapon_choices %in% c("Other", "Unclear from report"), "other", weapon_cat),
         weapon_cat = ifelse(weapon_choices %in% c("Toy weapon" , "Unarmed"), "unarmed", weapon_cat),
         weapon_cat = ifelse(weapon_choices %in% c("Vehicle"), "vehicle", weapon_cat))

# Check types & preview 
fl_shootings %>% select(weapon_choices, weapon_cat) %>% glimpse()

# Check unique combinations 
fl_shootings %>% distinct(weapon_choices, weapon_cat)
```

## 2.6. Add indicators 
```{r}
# Add race binary vars 
fl_shootings %<>% 
  mutate(black = ifelse(race=="Blk", 1, 0),
         white = ifelse(race=="Wh", 1, 0),
         other_race = ifelse(!race %in% c("Blk", "Wh"), 1, 0),
         other_race = ifelse(is.na(race), NA, other_race) )

# Check types & preview 
fl_shootings %>% select(race, white, black, other_race) %>% sample_n(30)

distinct(fl_shootings, race, black, white, other_race)
```


**********************************************************

# 3) SAVE DATA
```{r}
save(fl_shootings, file="../data/fl_shooting_data_clean.RData")
```
